tools/deploy/helm/vector/requirements.txt
boto3==1.22.0
pyyaml==5.3.1
